requirements.txt
flask
line-bot-sdk==2.4.0
apscheduler
